@@ -1,17 +1,17 @@
 module github.com/gardener/landscaper-service/test/integration
 
-go 1.18
+go 1.19
 
 require (
 	github.com/gardener/component-spec/bindings-go v0.0.66
 	github.com/gardener/landscaper-service v0.0.0-00010101000000-000000000000
-	github.com/gardener/landscaper/apis v0.45.0
-	github.com/gardener/landscaper/controller-utils v0.45.0
+	github.com/gardener/landscaper/apis v0.46.0
+	github.com/gardener/landscaper/controller-utils v0.46.0
 	github.com/gardener/landscapercli v0.20.0
-	k8s.io/api v0.24.3
-	k8s.io/apimachinery v0.24.3
-	k8s.io/client-go v0.24.3
-	sigs.k8s.io/controller-runtime v0.12.2
+	k8s.io/api v0.26.1
+	k8s.io/apimachinery v0.26.1
+	k8s.io/client-go v0.26.1
+	sigs.k8s.io/controller-runtime v0.14.0
 )
 
 require (
